+++ src/main.ts
@@ -1,8 +1,8 @@
 import type Messages from './messages'
 import type { Rule } from './rule'
-import type { LangTypes } from './types/lang'
-import type { RuleType } from './types/rule'
-import type { ValidatorOptions, VoidFunction } from './types/validator'
+import type { LangTypes } from './types'
+import type { RuleType } from './types'
+import type { ValidatorOptions, VoidFunction } from './types'
 import AsyncResolvers from './async-resolvers'
 import Errors from './errors'
 import Lang from './lang'
