--- src/types/lang.ts
@@ -1,51 +1,0 @@
-export type LangTypes =
-  | 'ar'
-  | 'az'
-  | 'be'
-  | 'bg'
-  | 'bs'
-  | 'ca'
-  | 'cs'
-  | 'cy'
-  | 'da'
-  | 'de'
-  | 'el'
-  | 'en'
-  | 'es'
-  | 'et'
-  | 'eu'
-  | 'fa'
-  | 'fi'
-  | 'fr'
-  | 'hr'
-  | 'hu'
-  | 'id'
-  | 'it'
-  | 'ja'
-  | 'ka'
-  | 'km'
-  | 'ko'
-  | 'lt'
-  | 'lv'
-  | 'mk'
-  | 'mn'
-  | 'ms'
-  | 'nb_NO'
-  | 'nl'
-  | 'pl'
-  | 'pt'
-  | 'pt_BR'
-  | 'ro'
-  | 'ru'
-  | 'se'
-  | 'sl'
-  | 'sq'
-  | 'sr'
-  | 'sv'
-  | 'tr'
-  | 'ua'
-  | 'uk'
-  | 'uz'
-  | 'vi'
-  | 'zh'
-  | 'zh_TW'
